Flask-app/requirements.txt
flask
flask_cors
python-dotenv
bcrypt